*fixed styles (moved into the file). renamed folder, minor fixes
