+++ 6/src/actions/setUsername.js
@@ -1,4 +1,4 @@
-import  USERNAME from "../constants&types/types"
+import  USERNAME from "../constantsAndTypes/types"
 
 export default function setUserName (username) {
     return {
