+++ 6/src/common/AppHeader.js
@@ -10,14 +10,9 @@ import IconMenu from 'material-ui/IconMenu';
 import MenuItem from 'material-ui/MenuItem';
 import MoreVertIcon from 'material-ui/svg-icons/navigation/more-vert';
 
-import setUserName from "../actions/setUserName"
+import setUserName from "../actions/setUserName";
+import {styles} from "../styles";
 
-const styles = {
-    title: {
-        cursor: 'pointer',
-        textAlign: "center"
-    },
-};
 
 class AppHeader extends Component {
     constructor(props) {
