+++ 6/src/reducers/orderByReducer.js
@@ -1,4 +1,4 @@
-import { ORDER_BY } from "../constants&types/types";
+import { ORDER_BY } from "../constantsAndTypes/types";
 
 export const orderByReducer = (state = "", action) => {
     switch(action.type){
