+++ 6/src/styles.js
@@ -1,0 +1,37 @@
+ export const styles = {
+    title: {
+        cursor: 'pointer',
+        textAlign: "center"
+    },
+};
+
+export const posterSmallStyles = {
+    image: {
+        display: 'block',
+        float: 'left',
+        height: '400px',
+        width: '450px',
+        alt: "rocket"
+    },
+    container:{
+        padding: '10px',
+        height: '450px', 
+        borderSize: '1px',
+        borderStyle: 'solid'
+    },
+    text: {
+        fontSize:'25px',
+        padding: "10px",
+        display: "block"
+    }
+}
+
+export const autoCompleteStyle = {
+    div: {
+        float: "left",
+        width: "50%"
+    },
+    autoComplete: {
+        width: "350px"
+    }
+}
